refactor(ui): split runtime utility domains
Extract trace normalization, route diagnostics, and budget helpers from the large runtime-utils module. Keep runtime-utils focused on generic formatting and call-shape helpers while preserving the existing exported behavior.
